requirements.txt
beautifulsoup4==4.12.3
langchain==0.2.7
llama_index==0.10.55
openai==1.35.13
pinecone_client==2.2.4
pydantic==2.8.2
Requests==2.32.3
streamlit==1.36.0
